% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_quantiles.R
\name{categorize_by_quantiles}
\alias{categorize_by_quantiles}
\title{Categorize a Numeric Variable into Quantile-Based Categories}
\usage{
categorize_by_quantiles(
  numerical_data,
  numerical_var_name,
  categorical_labels,
  num_categories
)
}
\arguments{
\item{numerical_data}{A data frame or tibble containing the numerical column to categorize.}

\item{numerical_var_name}{A character string of the numerical column name to categorize.}

\item{categorical_labels}{A character vector of labels for each category. Note: it should match the number of categories specified.}

\item{num_categories}{An integer specifying the number of categories to split into. Must be greater than 1.}
}
\value{
A data frame or tibble with an additional categorical column based on quantiles.
}
\description{
This function takes a numeric column from a data frame or tibble and divides it into 2 or more of categories
by using a specified number of quantiles.
A new column with the suffix "_category" is added with the specified labels.
The number of categories and the label for each category are decided by users.
}
\examples{
# Example1: Using the penguins dataset to categorize body_mass_g
penguins_categorized <- categorize_by_quantiles(palmerpenguins::penguins, "body_mass_g", c("light", "average", "heavy"), 3)
# Move the new column body_mass_g_category to the front
penguins_categorized <- dplyr::select(penguins_categorized, body_mass_g_category, dplyr::everything())
# Print the result
head(penguins_categorized)
print(penguins_categorized)

# Example2: Using a simple tibble to categorize "score" into two categories
# Define a simple tibble
simple_data_1 <- tibble::tibble(id = c("1", "2", "3", "4", "5", "6", "7", "8", "9","10"), score = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 90))
data_categorized_1 <- categorize_by_quantiles(simple_data_1, "score", c("low", "high"), 2)
# Print the result
print(data_categorized_1)

# Example3: Apply categorize_by_quantiles to categorize "score" into four categories
# Define a simple tibble
simple_data_2 <- tibble::tibble(id = c("1", "2", "3", "4", "5", "6", "7", "8", "9","10"), score = c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9))
data_categorized_2 <- categorize_by_quantiles(simple_data_2, "score", c("low", "medium", "high", "very high"), 4)
# Print the result
print(data_categorized_2)
}
